+++ CHANGELOG.md
@@ -13,6 +13,7 @@ HumHub Changelog
 - Fix #7443: Exclude inactive users from `followers` and `following` lists
 - Fix #7447: Update the comment counter when deleting a comment
 - Fix #7456: Fix setting manager backwards compatibility
+- Enh #7468: Remove deprecated widget `DataSaved`
 
 1.17.2 (Unreleased)
 ----------------------